requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
scikit-learn==1.3.2
joblib==1.3.2
prophet==1.1.5
requests==2.31.0
pydantic==2.5.0
bcrypt==4.0.1
python-jose[cryptography]==3.3.0